refactor: utils for clone metadata of a post or topic
move clone functions from
`src/containers/Home.js` to `src/utils/clone-entity.js`
